Multi-site Dashboard Emails: Show prorated price in footer summary
